Proposal to Speak from Kiley Williams
I am really excited to speak to the DevOpsDC crowd. This would be awesome! Our team has worked really hard to make sure that security (what we call DevSecOps) makes the lives of all stakeholders as easy as possible when it comes to security. I would be happy to give you a 5-minute overview of what it is we do, and how we do it. Then you can determine if it is a fit for DevOpsDC.